utils/requirements.txt
google-auth==2.40.3
google-auth-oauthlib==1.2.2
google-auth-httplib2==0.2.0
google-api-python-client==2.175.0
psycopg2-binary==2.9.10
python-dotenv==1.1.1
openai==1.69.0